feat: add dynamic chain handling and improve API functions
- Introduced dynamic chain support in package API with `emit_package_init_from_dynamic`.
- Added `c4rt_datum_preload_arguments` and related functions for argument preloading in datum.
- Enhanced handling for LLVM IR and GC allocation in `c4rt2c`.
- Replaced `mimalloc` with `BDWgc` in build configuration.

+++ src/c4rt2/src/c4rt2/package.1.h
@@ -188,7 +188,7 @@ C4RT_IMPL void
 c4rt_package_init_from_dynamic_v1(struct c4_package_v1_t* pkg,
                                   c4_datum_t datum,
                                   const struct c4_package_v1_t** fn_data,
-                                  uint16_t fn_data_sz_bytes);
+                                  uint32_t fn_data_sz_bytes);
 
 /**
  * c4rt_package_init_from_result_v1(*pkg, datum) --
